Organize fs watcher code
