+++ crates/zng-ext-fs-watcher/src/service/read_to_var.rs
@@ -1,0 +1,109 @@
+use std::{
+    path::{Path, PathBuf},
+    sync::{
+        Arc,
+        atomic::{AtomicBool, Ordering},
+    },
+};
+
+use parking_lot::Mutex;
+use path_absolutize::Absolutize as _;
+use zng_clone_move::clmv;
+use zng_task as task;
+use zng_var::{Var, VarValue, var};
+
+use crate::{FsChangesArgs, WatcherHandle};
+
+pub struct ReadToVar {
+    read: Box<dyn Fn(&Arc<AtomicBool>, &WatcherHandle, ReadEvent) + Send + Sync>,
+    pending: Arc<AtomicBool>,
+    handle: WatcherHandle,
+}
+impl ReadToVar {
+    pub fn new<O: VarValue, R: 'static>(
+        handle: WatcherHandle,
+        mut path: PathBuf,
+        init: O,
+        load: fn(&Path) -> R,
+        read: impl FnMut(R) -> Option<O> + Send + 'static,
+        on_modify: impl Fn() + Send + Sync + 'static,
+    ) -> (Self, Var<O>) {
+        if let Ok(p) = path.absolutize() {
+            path = p.into_owned();
+        }
+        let path = Arc::new(path);
+        let var = var(init);
+        let on_modify = Arc::new(on_modify);
+
+        let pending = Arc::new(AtomicBool::new(false));
+        let read = Arc::new(Mutex::new(read));
+        let wk_var = var.downgrade();
+
+        // read task "drains" pending, drops handle if the var is dropped.
+        let read = Box::new(move |pending: &Arc<AtomicBool>, handle: &WatcherHandle, ev: ReadEvent| {
+            if wk_var.strong_count() == 0 {
+                handle.clone().force_drop();
+                return;
+            }
+
+            let spawn = match ev {
+                ReadEvent::Update => false,
+                ReadEvent::Event(args) => !pending.load(Ordering::Relaxed) && args.events_for_path(&path).next().is_some(),
+                ReadEvent::Init => true,
+            };
+
+            if !spawn {
+                return;
+            }
+
+            pending.store(true, Ordering::Relaxed);
+            if read.try_lock().is_none() {
+                // another task already running.
+                return;
+            }
+            task::spawn_wait(clmv!(read, wk_var, path, handle, pending, on_modify, || {
+                let mut read = read.lock();
+                while pending.swap(false, Ordering::Relaxed) {
+                    if let Some(update) = read(load(path.as_path())) {
+                        if let Some(var) = wk_var.upgrade() {
+                            var.modify(clmv!(on_modify, |vm| {
+                                vm.set(update);
+                                on_modify();
+                            }));
+                        } else {
+                            // var dropped
+                            handle.force_drop();
+                            break;
+                        }
+                    }
+                }
+            }));
+        });
+        read(&pending, &handle, ReadEvent::Init);
+
+        (Self { read, pending, handle }, var.read_only())
+    }
+
+    /// Match the event and flag variable update.
+    ///
+    /// Returns if the variable is still alive.
+    pub fn on_event(&mut self, args: &FsChangesArgs) -> bool {
+        if !self.handle.is_dropped() {
+            (self.read)(&self.pending, &self.handle, ReadEvent::Event(args));
+        }
+        !self.handle.is_dropped()
+    }
+
+    /// Returns if the variable is still alive.
+    pub fn retain(&mut self) -> bool {
+        if !self.handle.is_dropped() {
+            (self.read)(&self.pending, &self.handle, ReadEvent::Update);
+        }
+        !self.handle.is_dropped()
+    }
+}
+enum ReadEvent<'a> {
+    Update,
+    Event(&'a FsChangesArgs),
+    Init,
+}
